legacy/OpenManus/requirements-lock.txt
# Locked dependencies for OpenManus
# Generated: 2026-01-21
# Purpose: Ensure version consistency to prevent Browser initialization issues

# Core AI/ML
openai==1.55.0
anthropic==0.39.0
google-generativeai==0.8.3
groq==0.11.0
instructor==1.6.4
litellm==1.51.3
tenacity==9.0.0
tiktoken==0.8.0

# Browser Automation - CRITICAL VERSION LOCK
browser-use==0.1.40  # DO NOT UPGRADE without testing Browser.__init__()
playwright==1.48.0   # Compatible with browser-use 0.1.40

# Web Framework
fastapi==0.115.5
uvicorn==0.32.1
websockets==14.1
python-multipart==0.0.17
httpx==0.28.0

# Data Processing
pydantic==2.10.3
pandas==2.2.3
numpy==1.26.4

# Database
asyncpg==0.30.0
redis==5.2.0

# Utils
python-dotenv==1.0.1
loguru==0.7.3
rich==13.9.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Development
black==24.10.0
isort==5.13.2
mypy==1.13.0
pre-commit==4.0.1